add batchReader
This was needed because we wanted to allow a batch to have an empty
list of samples, and as we were taking the batchName from the first
sample, we couldn't build a VariantSource from an empty list.

    String batchName = samples.get(0).getBatch();

would fail. The solution is to read not a bare list of samples, but
a bean DbsnpBatch{ batchId, batchName, List<Sample> }.